+++ src/client.test.ts
@@ -1,8 +1,7 @@
 import { it, expect, afterEach, beforeAll, afterAll, describe } from "vitest";
-import { render, screen, cleanup } from "@testing-library/react";
 import { http, HttpResponse, type RequestHandler } from "msw";
 import { setupServer } from "msw/node";
-import { User } from "./User";
+import { fetchUser } from "./client";
 
 describe("Fetch from api", () => {
   const handlers: RequestHandler[] = [
@@ -13,7 +12,7 @@ describe("Fetch from api", () => {
           return HttpResponse.json({ id: 1, name: "John Doe", username: "jd" });
         }
         return new HttpResponse("Not found", { status: 404 });
-      },
+      }
     ),
   ];
 
@@ -23,31 +22,20 @@ describe("Fetch from api", () => {
     server.listen();
   });
   afterEach(() => {
-    cleanup();
     server.resetHandlers();
   });
   afterAll(() => {
     server.close();
   });
-
-  it("User ID:1, John Doe, jd", async () => {
-    render(<User id="1" />);
-
-    // check loading text
-    expect(screen.getByText("Loading...")).not.toBeUndefined();
-
-    // show user info
-    expect(await screen.findByText("John Doe")).not.toBeUndefined();
-    expect(await screen.findByText("jd")).not.toBeUndefined();
+  it("fetchUser return user data", async () => {
+    const result = await fetchUser("1");
+    expect(result).toEqual({
+      id: 1,
+      name: "John Doe",
+      username: "jd",
+    });
   });
-
-  it("User ID:2, Not found error", async () => {
-    render(<User id="2" />);
-
-    // check loading text
-    expect(screen.getByText("Loading...")).not.toBeUndefined();
-
-    // show not found
-    expect(await screen.findByText("Not Found")).not.toBeUndefined();
+  it("fetchUser throws error", async () => {
+    await expect(() => fetchUser("2")).rejects.toThrowError(/fetching error/);
   });
 });
