+++ src/types.ts
@@ -1,0 +1,5 @@
+export type UserModel = {
+  id: number;
+  name: string;
+  username: string;
+};
